Refactoring to address crash
This refactoring introduces a backwards incompatible API change: the `alwaysShowTouches` and `showTouchesWhenKeyboardShown` properties are now gone, and the semantics of `showTouches` has changed.

`showTouches` is now more predictable and reflects whether touches are being rendered or not. `alwaysShowTouches` is sort of replaced by `automaticallyManageTouchesWhenScreenMirrored`, which, if YES, will automatically set `showTouches`.

Added a notification (QTouchposeTouchesVisibleDidChange) for monitoring changes to `showTouches`.

This resolves #16 and resolves #11.